moving workbench-util to broadinstitute/workbench-libs
