+++ core/src/main/scala/org/broadinstitute/dsde/rawls/dataaccess/HttpExecutionServiceDAO.scala
@@ -6,9 +6,10 @@ import com.typesafe.scalalogging.LazyLogging
 import org.broadinstitute.dsde.rawls.metrics.RawlsExpansion._
 import org.broadinstitute.dsde.rawls.metrics.RawlsInstrumented
 import org.broadinstitute.dsde.rawls.model.ExecutionJsonSupport._
-import org.broadinstitute.dsde.rawls.model._
-import org.broadinstitute.dsde.rawls.util.SprayClientUtils._
 import org.broadinstitute.dsde.rawls.util.{FutureSupport, Retry}
+
+import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.duration.FiniteDuration
 import spray.client.pipelining._
 import spray.http.{BodyPart, MultipartFormData}
 import spray.httpx.SprayJsonSupport._
