+++ core/src/main/scala/org/broadinstitute/dsde/rawls/statistics/StatisticsService.scala
@@ -9,7 +9,8 @@ import org.broadinstitute.dsde.rawls.model.StatisticsJsonSupport._
 
 import org.broadinstitute.dsde.rawls.model._
 import org.broadinstitute.dsde.rawls.statistics.StatisticsService._
-import org.broadinstitute.dsde.rawls.util.{RoleSupport, FutureSupport, UserWiths}
+import org.broadinstitute.dsde.workbench.util.FutureSupport
+import org.broadinstitute.dsde.rawls.util.{RoleSupport, UserWiths}
 import org.broadinstitute.dsde.rawls.webservice.PerRequest.{PerRequestMessage, RequestComplete}
 import org.joda.time.DateTime
 import spray.http.StatusCodes
